Add proxies for indicators
This works for *some* lamps - however there are additional controllers for lamps that can read from MU that need more than just ports.

This also includes some port metadata stuff that comes up in components.  We should be able to eventually use this to generate a filtered list of available ports from which to select providing a custom editor for port ids to make it simpler for users.